server/backend/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2.20
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
httplib2==0.22.0
idna==3.10
oauthlib==3.2.2
pillow==10.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.9.0
pyparsing==3.1.4
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
shortuuid==1.0.13
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
